+++ packages/thirdweb/src/extensions/erc1155/drop1155.test.ts
@@ -7,7 +7,7 @@ import {
   TEST_ACCOUNT_B,
 } from "../../../test/src/test-wallets.js";
 import { type ThirdwebContract, getContract } from "../../contract/contract.js";
-import { sendAndConfirmTransaction } from "../../exports/transaction.js";
+import { sendAndConfirmTransaction } from "../../transaction/actions/send-and-confirm-transaction.js";
 import { getContractMetadata } from "../common/read/getContractMetadata.js";
 import { deployERC1155Contract } from "../prebuilts/deploy-erc1155.js";
 import { balanceOf } from "./__generated__/IERC1155/read/balanceOf.js";
