+++ packages/thirdweb/src/extensions/erc721/drop721.test.ts
@@ -7,7 +7,7 @@ import {
   TEST_ACCOUNT_B,
 } from "../../../test/src/test-wallets.js";
 import { type ThirdwebContract, getContract } from "../../contract/contract.js";
-import { sendAndConfirmTransaction } from "../../exports/transaction.js";
+import { sendAndConfirmTransaction } from "../../transaction/actions/send-and-confirm-transaction.js";
 import { getContractMetadata } from "../common/read/getContractMetadata.js";
 import { deployERC721Contract } from "../prebuilts/deploy-erc721.js";
 import { balanceOf } from "./__generated__/IERC721A/read/balanceOf.js";
