+++ packages/thirdweb/src/react/core/hooks/contract/useSendBatchTransaction.ts
@@ -1,5 +1,5 @@
 import { type UseMutationResult, useMutation } from "@tanstack/react-query";
-import { sendBatchTransaction } from "../../../../exports/transaction.js";
+import { sendBatchTransaction } from "../../../../transaction/actions/send-batch-transaction.js";
 import type { WaitForReceiptOptions } from "../../../../transaction/actions/wait-for-tx-receipt.js";
 import type { PreparedTransaction } from "../../../../transaction/prepare-transaction.js";
 import { useActiveAccount } from "../wallets/wallet-hooks.js";
