+++ packages/thirdweb/src/utils/extensions/drops/fetch-proofs-for-claimers.ts
@@ -98,7 +98,7 @@ export async function fetchProofsForClaimer(options: {
     }
     const [{ getContract }, { decimals: getDecimals }] = await Promise.all([
       import("../../../contract/contract.js"),
-      import("../../../exports/extensions/erc20.js"),
+      import("../../../extensions/erc20/read/decimals.js"),
     ]);
     const currencyContract = getContract({
       address: currencyAddress,
